More organism implementations, slight formatting changes, enlarged the main window.
Implemented the Dandelion and Wolf classes. Fixed Grass definition according to specification. Introduced general behavior classes (AnimalBehavior, PlantBehavior) and specific behavior classes for all organisms. Changed the position of some braces to be more consistent with the rest of the codebase. The main window is larger now (1200x800 instead of 900x600).